Refactor options subscription without outside component detection
reg-sub-raw avoids the issues with the input signals function having to
react on state changes, which doesn't work, because the signals are
determined the a subscription comes into existence (either for the first
time or after having been dropped).

For this the argument for :heraldry.state/options also is reduced to the
path, and all the outside enriching of the context isn't necessary
anymore.