Move options-subscriptions computation into its own subscription
This decomplects ::options a bit, because now there's a subscription
that simply provides all the data ::options needs to compute the actual
options.

It also improves performance slightly, because now that work doesn't
have to be redone in ::options, if any of the required input
subscriptions change, without component-type or entity-type changing.