Start migrating towards strings in JSON format
This should work much better with crowdin and reduce string duplication.
It'll also naturally provide a context for strings that are the same in
English, e.g. "Fess" as ordinary and "Fess" as point, avoiding the
current hacks with "Fess [ordinary]" strings, which also cause warnings
in PO editors and the translation platform.

Another advantage will be the use of namespaced keywords for this,
avoiding the cumbersome `string` in many cases. While still being able
to use it whenever an untranslated string needs to be assembled
manually.