Add size-only flag to aws s3 sync
This will avoid re-upload if the file size doesn't change. Currently the
check on mtime is not very useful, as the local time of last
modification never will match the upload mtime on S3.

A better/safer mechanism would be to hash the file content and store it
as e-tag. These tickets discuss such a feature, but it isn't implemented
yet:
- https://github.com/aws/aws-cli/issues/599
- https://github.com/aws/aws-cli/pull/575