Add a hook to change render-options scope on helm/ornament changes
If the scope is set to coat-of-arms, then any changes to helms and
ornaments remain invisible, which is confusing.
Similarly if it is set to coat-of-arms-and-helm and an ornament change
is done.

With this change the scope will adjust to the lowest level that is
required to see the change that was effected.