Remove reitit path parameter coercion
This caused problems with paths where numeric strings were expected but
others were given, which can't sensibly be handled right now, it seems:
https://github.com/metosin/reitit/issues/334

This attempt keeps all path bits strings and the respective libraries
inspect the version parameter to determine whether it is a valid integer
version or not (or nil, which would be fine).

In addition to that the not-found page was unified into a single
component, which later can be improved.