Favour no-component field parsing, effectively making components non-greedy
As the test case describes, this makes sure "per pale azure and argent,
a fess sable" (and others like it) will parse the ordinary as part of
the per pale field, rather than as a component of the argent subfield.

Parentheses can be used to include the components in a subfield, or to
make it explicit that "(argent)" has no components.