Simplify ordinary parsing
This makes the single ordinary rule a bit more complex, but it allows
more fine-grained control over the allowed options (and lines/cottising)
for each ordinary type. It also avoids the many intermediate rules of
ordinar-type-* and ordinary-*.

A slight disadvantage is that ast->hdn now needs to recognize the
ordinary-type and ordinary-option nodes by checking for the actual leaf
node, as there is no wrapping rule anymore. In the case of ordinary-type
this ought to be the list of types known to the code anyway, with a
match of the keyword (albeit uppercased), so that is fine. In the case
of ordinary-option the function needs to know which leaf tags to expect
and how to treat them anyway, so it isn't a problem either.