Simplify partition parsing
This makes the single partition rule a bit more complex, but it allows
more fine-grained control over the allowed options (and lines/fields)
for each partition type. It also avoids the many intermediate rules of
partition-type-* and partition-*.

A slight disadvantage is that ast->hdn now needs to recognize the
partition-type nodes by checking for the actual leaf node, as there is
no wrapping rule anymore. In the case of partition-type this ought to be
the list of types known to the code anyway, with a match of the
keyword (albeit uppercased), so that is fine