Filter charge-types for terminals known to the parser
There are some dodgily named charges, as they are user created. It's
hard to prevent shenanigans or unexpected results entirely, but
filtering the dynamic list of charge-types for strings that the parser
can determine are "bad" charge types should take care of the fast
majority of cases where this is a problem.