Add parsing of partition field references
The transform step will try to take something like "i. and 4th" and set
the first field to the actual field described, the 4th to a reference of
the first. That way they'll remain locked even when being edited in the
interface.

A special case is a reference that would be set for any of the mandatory
fields, e.g. "i. and ii.", there shouldn't be any reference fields
there, so they always use the explicit field data.

Another special case is a default reference field, e.g. the fourth of a
quartered partition that references the first. If the referenced field
also is a field reference, then that reference is used for the fourth
field, rather than the initial reference to the first.
This shouldn't be possible right now because of the special case
above... all default references should point at one of the mandatory
fields, which must not be field references anyway, but default fields
might change in the future and with multiple base fields it is
conceivable that this situation could arise.