Generate bad-charge-types once
Doing this by parsing each charge type actually is pretty expensive,
resulting in a second lag for the user when the parser with custom
charge types is generated.

This way it is done once when the parser is generated, and all the
suggestions are collected in a set for easy and fast look up later.