Add own test reporter for better diff and a bit of colour
heraldry.test uses the minimal setup to run the tests in the same way
shadow.test.node does. The call of `get-test-data` MUST be done inside
this namespace, as shadow-cljs magically adds all test namespaces as
dependencies of the test runner namespace. This is also the reason for
the metadata on the namespace and `reset-test-data!`.
See issue 1005 on shadow-cljs.

ams-clj/clansi only supports Clojure, I pulled the relevant data and
functions into a ClojureScript namespace, removing macros I didn't need.
It's licensed under MIT.