Add spec requires
This moves the required spec namespaces back into the spec namespaces
where they are needed. That is, any namespace referring to a spec from
another must require that other namespace to make sure it is known at
compile time.

The exception is `spec.heraldry.field`, which itself requires many
other namespaces, so it'd result in a dependency cycle. For that
special case `su/spec` wraps around `:heraldry/field`.