Add anchor-choices and orientation-choices
This avoids a lot of repetition of the string constants for the
choices, it also normalizes the order of choices and adds choice
groups.

The separation of anchor and orientation is a bit cleaner now, the
order of groups differs for both areas, as certain choices are more
likely/useful.

This also means the custom order of some options no longer apply.

Some options have been removed, because they weren't sensible:
- gyronny-n: anchor :angle
- tierced-per-pall: anchor :angle