Load dummy SVG saved by Adobe Illustrator to fix Inkscape bug
There seems to be a bug in Inkscape, which strips ENTITY definitions
from SVGs saved with Adobe Illustrator, but still keeps those ENTITY
references in the namespace definitions, breaking SVGO when it tries
to load such files after they were saved in Inkscape.

I couldn't find a proper fix for that at the moment, but what does
work is loading such an original AI file, as it stores the references
in some global repository inside SVGO, which bleeds into subsequent
parsing attempts. So loading a dummy AI file once in the beginning
will allow loading the "broken" Inkscape ones afterwards, hopefully
reducing the impact for users.

Finally, the xmlns aliases are important. And AI seems to use "x",
"i", and "graph" mainly. If others are observed in the wild, then they
should be added.