Simplify routes with better trailing slash handling
Whether trailing slashes are desired depends on the route definition.
If the route has it, but the path doesn't, then it still matches but
any path creation by route name will give the desired format.
If it doesn't have a trailing slash, then it works the same just
without the slash.