Move maintenance logic into frontend.maintenance
Move the branching of using the maintenance mode into router/view
instead writing the view function into the state in on-navigate.

This is much simpler and is easier to be refactored into a version
where the maintenance mode may change on the fly (e.g. by checking via
some API) and the view is updated automatically, rather than requiring
a route change.