Make saving transition smoother
The key thing is to return :done with the new entity in
::prepare-for-editing, so there's no split second in which the UI has
to render a "loading" page just because the new entity (which already
is available) hasn't been written to the target path yet.

In addition a new ::replace-data event and a ::replace-route side
effect clean up the save completion a bit.