Fix entity-for-editing loading in by-id-view
The previous approach caused a history state to be written right after
saving, because the subscription to the form data noticed a change in
id/version, forcing the original entity-for-editing to be written
again. The change to the route based on the new id/version then caused
the new version to be written, creating another state in the history.