Remove default entity data from initialize-db
This makes it harder to decide whether the form data ought to be
considered the initial state or not, which breaks the hdn-hash
loading.

With the new handling of by-id-view the change actually isn't
necessary anymore.

Finally, no need to wrap load-hdn in a go block, as it returns a go
block itself.