Fix fimbriation for charges without outline
If an SVG element has a fill, and a stroke of "none", then overwrite
that stroke colour anyway for fimbriation. This can't be done for all
stroke or fills of "none", because they might designate areas that
must not be filled, so filling them for fimbriation also would be
wrong.