Disable convertStyleToAttrs during SVG optimization
This doesn't seem to work correctly in all cases, so the patch
includes a function to apply a subset of style attributes to SVG
elements, rather than letting SVGO do it.

The backend version of SVGO actually disables this by default, which
resulted in different behaviour frontend-side and backend-site, where
an uploaded SVG had broken outlines in the frontend, but the SVG that
was optimized backend-side could be downloaded and would subsequently
work with an outline if uploaded again.