Strip clip-path references from charge SVGs on upload
This doesn't play well on Firefox and is likely to cause issues in the
interface. There might be SVGs that break due to this, but the ones
I've seen in the wild don't use the clip path for anything relevant.
This can be revisited if there show up cases where this approach
doesn't work.