WIP: render "other" charge data, include fimbriation
For now this pretty much reproduces the previous way the charges were
rendered, but without vertical-mask, positioning, rotation, etc.,
which already is done via the properties.

Fimbriation now takes the parent environment's width or height as
percentage-base, previously it was the size of the charge.
The scaling required for this during dilation doesn't seem to work
with the current "Coolshield Sword" charge in the staging DB.