Add strategy to overlap masks of subfields while rendering
Subfields with overlaps (by default all even fields, but each field
type can define that logic) are rendered first, slightly larger than
they need to be, then the other subfields are rendered on top of them,
so there aren't any anti-aliasing issues at the edges where they meet.

This only has an effect when masks are used, i.e. clips in the browser
will ignore the stroke width of the edge.

The remaining cases where this is results in a slight extra overlap
occur when there is an odd number of fields that meet in one point,
namely tierced-per-pall and gyronny-n.