Refactor subfield structure
Previously subfields were defined like so:
{:type ...
 :fields [{:type :heraldry.field.type/plain
           :tincture :or}
          {:type :heraldry.field.type/ref
           :index 0}]}

This was problematic, because environment, shape, etc. for [:fields 0]
served a double role, once to inform the subfield of its surrounding
environment/shape, and once for child components of the field, but
those are not equivalent.

The solution is to add another abstraction level of the subfield,
which also avoids the previously cumbersome workaround field type of
:ref. This is not an explicit subfield type:

{:type ...
 :fields [{:type :heraldry.subfield.type/field
           :field {:type :heraldry.field.type/plain
                   :tincture :or}
          {:type :heraldry.subfield.type/reference
           :index 0}]}

And subfield types have their own logic on how to refer to the parent
and parse out the subfield's environment and shape.