Fortify shape shrinking against endless recursion
paperjs-offset sometimes returns a path that has several closed paths,
trying to shrink that results in an endless loop, at least for the one
I've observed:

M219.64515,39.95564c0.70091,-1.8409 0.25565,-3.92179
-1.13723,-5.31466l-0.00001,-0.00001c-1.86123,-1.86123
-4.84674,-1.96075 -6.82778,-0.22759c-1.98104,1.73316 -2.27915,4.70541
-0.68176,6.79747c0,0 0,0 0,0.00001c1.00776,1.31985 2.60505,2.05324
4.26287,1.95731c0,0 0,0 0,0c1.96653,-0.1138 3.68299,-1.37162
4.3839,-3.21252zM228.43686,56.00161c13.25846,6.90842 29.98522,10.69526
56.7955,8.71279c-10.85462,31.00407 -35.48095,88.69764
-9.00504,194.78308c-3.18833,0.74395 -0.90069,-0.67907
-18.57804,6.1699c-15.73056,6.96715 -32.34096,18.80179
-45.08382,38.79184c-12.08163,-1.81664 -32.01985,-6.56267
-62.50421,13.24155c-33.86695,-24.94668 -51.78627,-16.61504
-62.95313,-13.75411c-12.87009,-21.57835 -29.02634,-31.31176
-44.75743,-38.27927c-17.61058,-6.8786 -15.41187,-5.49454
-18.60907,-6.24547c26.00321,-107.47173 1.63083,-164.42575
-8.97436,-194.71038c26.80554,1.97724 43.53601,-1.80115
56.79587,-8.7099c14.78919,-9.27235 14.10491,-10.44738
15.93589,-12.76461c19.6948,0.10021 36.55081,-7.46346
49.33675,-16.85141c13.5647,-11.04945 11.10549,-9.51127
13.21257,-11.65077c2.09773,2.13478 -0.45616,0.51134
13.11586,11.65077c12.78536,9.38769 29.58254,16.92676
49.24559,16.85018c0.62136,0.94898 0.10242,0.89248
2.73576,3.65751c2.22734,2.33872 6.21587,5.33319 13.29128,9.10832z

Grabbing the longest string from all the z-terminated strings inside
that path seems to work well enough.

It's also worth noting that this only happened together with another
bug, where the achievement's coat of arms was rendered with
target-width 300 (hence the large numbers in the path) and then
shrunken with steps of 5. With a coat-of-arms-target-width of 100,
introduced after this commit, the issue did not occur.

Still worth keeping the fix in general.