Fix small charge preview generation
Turns out there seems to be a bug in converting SVG to PNG for small
width/heights, with a width of 64 and lower for a thin charge like the
torse, this results in a PNG where part of the charge is cut off at
the bottom. The logic seems fine in the code, tho.

So just generate a slightly larger preview and use that for the charge
type choices.