Remove options-subscriptions system in favour of raw subscription
That system was far too complex in an attempt to feed all the required
subscriptions into an option subscription for a component, which
is much easier with a raw subscription, and that should be fine.

It also made the use of reaction-or-cache for options difficult, so
there was a server-side rendering impact.