Remove svg-namespaced tags and attributes from charge SVGs
These namespaces would be defined in the <svg> element, but that one
is turned into a <g> element, leaving namespaced tags and attributes
that are illegal inside SVGs generated with these charges.

Since those tags and attributes shouldn't be required at all for the
rendering, they can simply be stripped.