Fix bounding-box of bend ordinary
The bounding-box is in the parent space, so it has to be transformed
back into it before it can be used for infinity path calculations.

This bug became apparent in the 4th quarter of a quartered field,
because the bend's local bounding-box is horizontal and not very tall,
causing strange results in the 4th quarter.