Relax spec for entity.attribution
The :access field is outside of the attribution data, making it
difficult to enforce some constraints only if access is public.

There could be two versions, one for private and one for public
attribution, but that seems too much effort and complexity for little
gain.

For now just simplify the spec and keep the access validation in the
frontend and backend.