Add new mechanism to override render-options for collections
This removes the last render-options-paths starting with :context,
making render-options look-ups path based.

A new feature is that the absense of :render-options-path in the
context simply defaults to an empty render-options map, which results
in default values for all options. This simplifies cases where the
render-options are the default anyway.

The override for collections differs slightly from the previous
implementation, because it now overrides ONLY if the value specified
in the collection render-options is not nil or :none, so in order to
override things like :outline?, the checkbox has to be manually turned
on/off in the collections render-options.

Sensible defaults are set in defaults/collection-data for new
collections.