Add auto anchor for fess ordinary
Starting with the fess ordinary, groups of ordinaries of the same type
and the anchor :auto should be positioned and resized based on their
number and order, leaving a certain margin and taking line heights and
cottise heights into account.

For the line height a line length of field-width is assumed, and the
default for line styles of auto positioned bars is set to :field-width
if there are at least 2 of them, otherwise the single auto positioned
bar simply will use :fess and behave as before.

The ordinaries still can use :offset-y to adjust their relative
position and override their :size and line's :size-reference.