Improve shape shrinking algorithm
paperjs-offset sometimes generates multiple closed paths, with one (or
perhaps even more?) of them being a glitchy, small shape at one of the
original shape's corners. Previously the "real" shape was detected by
the length of its string representation, under the assumption that
it'd be a more complex shape.

This is not sufficient, as it results in the wrong shape to be picked
in some cases, so now actually inspect the shapes' areas and pick the
largest, which should be the one we want.