Add frontend.canvas for bounding-box calculation of SVG data
This renders SVG data to a canvas and inspects the painted pixels to
determine the approximate bounding box of an SVG.
It's not perfectly precise, but the algorithm should guarantee that
the SVG data does not extend past the bounding box.

If an SVG has a viewbox that already is the bounding box (initially
passed to the function as the starting point), then the detected
bounding box is only used if it is actually smaller than the
viewbox, which takes care of cases where the viewbox is already the
perfect bounding box, but the algorithm introduces inaccuracies and
makes it slightly bigger.