Improve charge shape algorithm
First go along the corners of the pixels identified as edges, rather
than using the middles of the square pixel sides, which will be more
accurate.

Then simplify the shape path with the Ramer-Douglas-Peucker algorithm
to make it smaller.

Also allow passing a draw-svg function to functions requiring to turn
SVG into pixels, so the backend can provide its own function for the
migration or future operations.