Tweak escutcheon shapes and change their bounding boxes
This changes the effective environment bounding boxes for a lot of
escutcheons, improving the default alignment of partitions and
ordinaries such as bends and saltires.

This will change existing arms substantially, and fine tuning of the
user to position ordinaries and charges likely won't match anymore
what the user intended.