Change field order for vairy/potenty/lozengy/masony/papellony
Some of these field partitions/variations had inconsistent rendering
order of the first and second tincture (or more generally field).

The rule adhered to now is that the first tincture (historically
usually the metal for these variants) is the one touching the top or
left of the surrounding field, at least in the default rendering,
offsetting, stretching, rotation, etc. can change this, but that can't
be helped.

This is a breaking change for some coats of arms using any of these
field partitions, but the effort of a full migration probably isn't
justified.

Fixes #7