Add tooltip to "inherit environment", shorten the text
The example in the checkbox text makes it very wide, at times reaching
into the "Load arms" link next to it. Changing the text won't improve
it for all cases, but I couldn't find a good fix that always works,
and the text was too long anyway, so this seems like a sufficient
solution.