Rewrite list pages served for crawler bots
This now limits the page to default-page-size (20) and has a "next
page" link for the crawler to find the next 10.

This hopefully also will improve the memory leak situation in the
backend, I suspect it was due to the large queries and the bug that
made the actual rendering crash, because I forgot to update this after
introducing the new search-filter namespace, it still tried to do the
old list.